requirements.txt
beautifulsoup4==4.12.2
openpyxl==3.1.2
selenium==4.10.0
mysql-connector-python==8.0.33